requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
bcrypt==4.0.1
PyJWT==2.8.0
